refactor: move interpreter loop to own file
The interpreter loop is most likely going to be the largest function in
the entire project. To keep code tidy, mostly to not bloat other files,
this function is kept in its own file (and subsequently module).
